requirements.txt
Django==1.9.7
django_q==0.7.18
djangorestframework==3.5.4
